[FancyZones] Rename "fancyzones/tests" -> "fancyzones/FancyZonesTests" (#11492)
