--- src/modules/fancyzones/tests/UnitTests/pch.cpp
@@ -1,5 +1,0 @@
-// pch.cpp: source file corresponding to the pre-compiled header
-
-#include "pch.h"
-
-// When you are using pre-compiled headers, this source file is necessary for compilation to succeed.
